+++ functions/Export-MySQLiteDB.ps1
@@ -1,5 +1,3 @@
-
-
 Function Export-MySQLiteDB {
     [cmdletbinding(SupportsShouldProcess)]
     Param(